Allow concrete changes to be suggested by a PR
The force of github is its ability to facilitate dialogue about changes based on the change itself. 

Asking for inputs and feedback exclusively as a (well described) issue discards concrete suggestsions to the text in the form of a pull request. 

Pull requests facilitates debate and discussion based on each change and - by that - supports a more fruitful dialogue. Also, having dialogue about concrete changes in a pull request will make it easy to track down what dialogue went into a specific change to a text.

A good case is issue #157 where adding the word `acoustic` seams well documented but its not easy for the reader to see the change in context without searching for the document. With #157 suggested as a pull request it would be easy understand the implications for the whole paragraph. 

I suggest that the documentation on how to provide feedback is extended with information describing that concrete changes may be suggested by a pull request if prefered.