Home page left padding
The home page included padding of 160px to allow for the navbar. This is
unnecessary for this page since there's no navbar. Added a style to set
the padding to 15px when the body element comes immediately before the
container. The remaining pages have a navbar and should have the
padding.