Update of 83C - Section 5.1.2 for clarity
Original contained a mix of descriptive and background information on what is a holder of key assertion (etc) mixed in with the Requirements (shall and may statements).  This can be confusing for the readers to delineate the descriptive from the actual requirements. 

Suggested rearrangement of this sub-section.  No additional requirement specific changes made directly in this PR.  Issues opened for requirement specific changes.