A copy edit review - batch #1
Introduction General:
Switching between subject, subscriber, and claimant without clearly
aligning with the graphic. I did not change since it wasn’t clear which
was preferred. Suggest updating to be more consistent as part of the
next revision.
Section 1.2:
For example, in order to maintain integrity of the service, it may be
desirable to know the home ZIP Code of a user for purposes of census
taking or petitioning an elected official but where it is not necessary
or desirable to know the underlying identity of the person. – Awkward
example. Suggest shortening.
Section 2.2:
Real-live v. Real-world. They are used interchangeably. I changed all to
real-life for consistency.
The term validation is not used at all in IAL2 or IAL3. There is also no
discussion of evidence in either of these descriptions. Not sure if this
is intentional.
In the discussion of IAL 2&3 and anonymity there is the following
statement: “It may be beneficial to the user that uses a high assurance
CSP for transactions at lower assurance levels.  For example, a user may
maintain an IAL3 identity, yet should be able to use their CSP for IAL2
and IAL1 transactions.” This does not explain a benefit. I provided some
suggested re-wording.
Section 4:
We state that there is only one outcome of identity proofing here, but
in section one we list 4 outcomes. We should update and harmonize the
language in one of the two sections.
4.1.1.6 Address Confirmation
The term verified and validated seem to be used interchangeably with
regard to the address of record matching what is in records. I am not
sure if this is intentional or if one is preferred.
4.4.1.8 Security controls
Very confusing control. It says they shall have moderate or high and
then says they should have moderate.
Section 5
The table for identity evidence uses a wide array of terms for the
person being proofed – from applicant, to subject, to person, to
individual. I am not sure if this is intentional.

Left off at Table 5-2