EN's pre-werb edits
Line 189: is "physical representation of the underlying subject" common terminology? If not, recommend changing because it sort of sounds like we don't need to know what you look like. When in reality it's saying we don't need to know who you are, just that you're the same person who registered for this account, right? Maybe say something like, "may not mean that the real-life identity of the subject is known."

Lines 215, 219, 232, and 244 all say "This section contains both normative and informative material." Is there a reason for this as opposed to just saying it once before diving into each section (e.g., line 209)?

Line 223: I thought it was confusing to explicitly say in IAL2 and IAL3 that attributed can be asserted by CSPs to RPs, but not to say it in IAL1. A CSP could assert an attribute by a CSP to an RP, it would just be treated as a self-asserted attribute. So I added the parenthetical to the end of line 223: "Any attributes provided in conjunction with the authentication process are self-asserted or should be treated as such (including attribute a Credential Service Provider, or CSP, asserts to an RP)." Make sense?