Pre-werb edits_intro
Line 96: "even though some authenticators that are used remotely may be used for physical authentication." Does this mean virtual in-person? Add an example to clarify?

Line 96: So does "this revision of these guidelines" mean what will be the new version of -A, -B, etc., but not necessarily pertain to future revisions of each of those volumes?

Line 104: what's the difference between "direct correlation" and "fulfills the criteria"? Does that mean -63 might go beyond what's required in those standards?

Line 126: "Additional (minimum) technical requirements were specified for the CSP, protocols used to transport authentication information, and assertions if implemented within the digital identity model." In this sentence, are protocols and assertions examples of the requirements specified for the CSP? If so, change comma after CSP to a colon, because as is it looks like a list of 3 distinct things.